+++ ui/.env.development
@@ -10,16 +10,3 @@ VUE_APP_MOCK = true
 # API接口地址
 VUE_APP_API_SERVER = http://127.0.0.1:8085/api/v1
 VUE_APP_API_AGENT =  http://127.0.0.1:8086/api/v1
-
-# local
-VUE_APP_API_SERVER = http://192.168.60.103:8085/api/v1
-VUE_APP_API_AGENT =  http://192.168.60.103:8086/api/v1
-
-# 测试开发环境
-# VUE_APP_API_SERVER = http://123.249.104.232:8085/api/v1
-# VUE_APP_API_AGENT =  http://123.249.104.232:8086/api/v1
-
-
-# 测试开发环境
-#VUE_APP_API_SERVER = http://123.249.104.232/api/v1
-#VUE_APP_API_AGENT =  http://123.249.104.232/api/v1
